+++ R/utilities.R
@@ -15,14 +15,14 @@ calculate_qvalue <- function(pvals) {
     if (length(pvals) == 0)
         return(numeric(0))
 
-    qobj <- tryCatch(qvalue::qvalue(pvals, lambda=0.05, pi0.method="bootstrap"), error=function(e) NULL)
+    qobj <- tryCatch(qvalue::qvalue(pvals), error=function(e) NULL)
 
     if (inherits(qobj, "qvalue")) {
-        qvalues <- qobj$qvalues
+        qv <- qobj$qvalues
     } else {
-        qvalues <- NA
+        qv <- rep(NA, length(pvals))
     }
-    return(qvalues)
+    return(qv)
 }
 
 get_geneSet_index <- function(geneSets, minGSSize, maxGSSize) {